Folder_CRUD_2

+++ app/Http/Controllers/Api/FolderController.php
@@ -7,6 +7,7 @@ use App\Models\Folder;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class FolderController extends Controller
@@ -33,10 +34,7 @@ class FolderController extends Controller
     }
 
     public function editFolder($id ,Request $request){
-        $user = User::find(Auth::user()->id);
-        if(!$user){
-            return response()->json(['error' => 'User not found'], 404);
-        }else{
+            $user = User::find(Auth::user()->id);
             $validator = Validator::make($request->all(), [
                 'name' => 'required',
             ]);
@@ -53,17 +51,98 @@ class FolderController extends Controller
                     return response()->json(['message'=>"unauthorized"], 401);
                 }
             }
-        }
+        
     }
     
+    public function addFilesToFolder($id, Request $request){
+        $user = User::find(Auth::user()->id);
+        $folder = Folder::find($id);
+        $validator = Validator::make($request->all(), [
+            'files' => 'required|array',
+            'files.*' => 'required|file',
+        ]);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
+        }
+            $files = $request->file('files');
+            $totalSize = 0;
+
+            foreach ($files as $file) {
+                $totalSize += $file->getSize();
+            }
+            $totalSizeInGB = $totalSize / (1024 * 1024 * 1024);
+
+            $documentsSize = $user->getMedia('documents')->sum('size') / (1024 * 1024 * 1024);
+            $trashSize = $user->getMedia('trash')->sum('size') / (1024 * 1024 * 1024);
+
+            if (($totalSizeInGB + $documentsSize + $trashSize) > 1) {
+                return response()->json(['error' => 'Storage limit exceeded. Maximum allowed storage (docs and trash) is 1 GB.']);
+            }
+
+            
+            // $urls = collect($files)->map(function ($file) use ($folder) {
+            //     $media = $folder->addMedia($file)->toMediaCollection('documents');
+            //     return asset($media->getUrl());
+            // });
+
+            // foreach($files as $file){
+            //     $user->addMedia($file)->preservingOriginal()->toMediaCollection('documents');
+            // }
+
+            $urls = collect($files)->map(function ($file) use ($folder, $user) {
+                $media = $folder->addMedia($file)->toMediaCollection('documents');
+                $user->addMedia($file)->toMediaCollection('documents');
+                return asset($media->getUrl());
+            });
+
+            
+            return response()->json(['urls' => $urls]);
+        
+    }
+
+    public function showFilesInFolder($id)
+    {
+        $user = User::find(Auth::user()->id);
+        $folder = Folder::find($id);
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
+        } else {
+            $files = $folder->getMedia('documents');
+
+            if ($files->isEmpty()) {
+                return response()->json(['message' => 'No files', 'files' => []]);
+            }
+        }
+        $fileData = $files->map(function ($file) {
+            return [
+                'id' => $file->id,
+                'url' => asset($file->getUrl())
+            ];
+        });
+        return response()->json(['files' => $fileData]);
+    }
+
     public function deleteFolders(Request $request){
             $user = User::find(Auth::user()->id);
-            $folders = $request->input('folders');
-            foreach ($folders as $folder) {
-                if($folder){ 
-                    $folder->move($user, 'trashFolders');
-                    $folder->delete();
+            if (!$user) {
+                return response()->json(['error' => 'User not found'], 404);
+            } else {
+                $folders = $request->input('folders');
+                foreach ($folders as $folder) {
+                    $folder=Folder::find($folder);
+                    if ($folder) {
+                        $files = $folder->getMedia('documents');
+                        foreach ($files as $file) {
+                            $file = $folder->getMedia('documents')->find($file);
+                            if ($file) {
+                                $file->move($user, 'trash');
+                                $file->delete();
+                            }
+                        }
+                        $folder->delete();
+                    }
                 }
+                return response()->json(['message' => 'Folders moved to trash']);
             }
     }
 }
